Amazing commit. Using butterknife to cleanup the code. Replace all int to double. Now there is a working support for number higher than 9, because you can compose, for exaple, 451, typing 4, 5 and finally 1, in the calculator.

+++ app/src/main/jni/calculator_native.c
@@ -17,20 +17,20 @@
 #include <jni.h>
 #include <stdio.h>
 
-jint
-Java_it_stefanocappa_ndkcalculator_Calculator_sum( JNIEnv* env, jobject thiz, jint a, jint b ) {
-    jint result = a + b;
+jdouble
+Java_it_stefanocappa_ndkcalculator_Calculator_sum( JNIEnv* env, jobject thiz, jdouble a, jdouble b ) {
+    jdouble result = a + b;
     return result;
 }
 
-jint
-Java_it_stefanocappa_ndkcalculator_Calculator_mult( JNIEnv* env, jobject thiz, jint a, jint b ) {
-    jint result = a * b;
+jdouble
+Java_it_stefanocappa_ndkcalculator_Calculator_mult( JNIEnv* env, jobject thiz, jdouble a, jdouble b ) {
+    jdouble result = a * b;
     return result;
 }
 
-jint
-Java_it_stefanocappa_ndkcalculator_Calculator_div( JNIEnv* env, jobject thiz, jint a, jint b ) {
-    jint result = a / b;
+jdouble
+Java_it_stefanocappa_ndkcalculator_Calculator_div( JNIEnv* env, jobject thiz, jdouble a, jdouble b ) {
+    jdouble result = a / b;
     return result;
 }
